# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Sylvara Platform API
  version: 1.0.0
  description: |
    Public and portal APIs for the Sylvara Paper Group platform.
    Errors use a standard envelope: `{ "error": { "code", "message" } }`.
  contact:
    email: engineering@sylvara.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.sylvara.com
    description: Production
  - url: http://localhost:3000
    description: Local website (contact/health routes)
...
paths:
...
  /api/contact:
    post:
      tags: [Leads]
      summary: Website contact / quotation intake
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "200": { description: Accepted }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 15-minute HS256 access token; refresh via rotating refresh tokens.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
    LeadInput:
      type: object
      required: [name, email, message]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        company: { type: string }
        topic: { type: string }
        category: { type: string }
        message: { type: string, maxLength: 10000 }
        source:
          type: string
          enum: [website, whatsapp, telegram, event, api]
...